+++ client/package.json
@@ -62,7 +62,7 @@
       "src/**/*.{js,jsx}",
       "!src/index.js",
       "!src/registerServiceWorker.js",
-      "!src/testUtils.js"
+      "!src/test-helpers/"
     ],
     "snapshotSerializers": [
       "enzyme-to-json/serializer"
